Refactored to separate modules
