+++ app/articles/__init__.py
@@ -1,16 +1,19 @@
 from datetime import timedelta
-from typing import Optional, Union
 
-import requests
-from flask import abort, current_app, json
-from werkzeug.exceptions import NotFound, Unauthorized
+from flask import abort, current_app
 
 from app import get_current_locale
-from app.articles.pages import get_content
-from app.extensions import redis_client
 
+REQUEST_TIMEOUT = 5
+GC_ARTICLES_AUTH_API_ENDPOINT = "/wp-json/jwt-auth/v1/token"
+GC_ARTICLES_AUTH_TOKEN_CACHE_KEY = "gc-articles-bearer-token"
+GC_ARTICLES_AUTH_TOKEN_CACHE_TTL = int(timedelta(days=1).total_seconds())
+GC_ARTICLES_FALLBACK_CACHE_PREFIX = "gc-articles-fallback--"
+GC_ARTICLES_FALLBACK_CACHE_TTL = int(timedelta(days=7).total_seconds())
 GC_ARTICLES_CACHE_PREFIX = "gc-articles--"
+GC_ARTICLES_DEFAULT_CACHE_TTL = int(timedelta(days=1).total_seconds())
 GC_ARTICLES_NAV_CACHE_TTL = int(timedelta(days=5).total_seconds())
+
 
 def _get_alt_locale(locale):
     return "fr" if locale == "en" else "en"
@@ -19,42 +22,6 @@ def _get_alt_locale(locale):
 def set_active_nav_item(items=[], url="") -> None:
     for item in items:
         item["active"] = True if item["url"] == url else False
-
-
-def get_nav_items() -> Optional[list]:
-    # @todo add caching
-    locale = get_current_locale(current_app)
-    items = _get_nav_wp(locale)
-    return items
-
-
-def _get_nav_wp(locale: str) -> Optional[list]:
-    nav_url = "menus/v1/menus/notify-admin"
-    if locale == "fr":
-        nav_url = "menus/v1/menus/notify-admin-fr"
-
-    cache_key = f"{GC_ARTICLES_CACHE_PREFIX}{nav_url}"
-
-    cached = redis_client.get(cache_key)
-    if cached is not None:
-        current_app.logger.info(f"Cache hit: {cache_key}")
-        nav_response = json.loads(cached)
-    else:
-        nav_response = get_content(nav_url)
-        current_app.logger.info(f"Saving menu to cache: {cache_key}")
-        redis_client.set(cache_key, json.dumps(nav_response), ex=GC_ARTICLES_NAV_CACHE_TTL)
-
-    nav_items = None
-
-    if isinstance(nav_response, dict) and "items" in nav_response:
-        nav_items = []
-        for item in nav_response["items"]:
-            nav_items.append({k: item[k] for k in ("title", "url", "target", "description")})
-
-        # always append a link called "preview", so that 'check_path' will find it
-        nav_items.append({"title": "preview", "url": "/preview", "target": "", "description": ""})
-
-    return nav_items
 
 
 # Return URL path for the language switcher
